requirements.txt
pybullet>=3.2.0
groq>=0.14.0
python-dotenv>=1.0.0
numpy>=1.26.0
